+++ Framework/Direct3D9/Render/GUI/Controls/UI_Button.h
@@ -9,15 +9,15 @@
 class UI_Button
 {
 public:
-	UI_Button(PCHAR ButtonText, int x, int y, LPCoreFunction ButtonFunction, COLOR32 Color = CLR_TEXT_DEFAULT);
+	UI_Button(PCHAR ButtonText, float x, float y, LPCoreFunction ButtonFunction, COLOR32 Color = CLR_TEXT_DEFAULT);
 	~UI_Button() { }
 
 public:
-	void Draw(int x, int y, bool Visible = true);
+	void Draw(float x, float y, bool Visible = true);
 
 public:
 	CoreString Text;
-	int X, Y;
+	float X, Y;
 	LPCoreFunction Function;
 	COLOR32 UnfocusedColor = CLR_BUTTON_UNFOCUSED;
 	COLOR32 FocusedColor = CLR_BUTTON_FOCUSED;
