+++ Framework/Direct3D9/Render/GUI/Controls/UI_GroupBox.h
@@ -9,15 +9,15 @@
 class UI_GroupBox
 {
 public:
-	UI_GroupBox(PCHAR Caption, int x, int y, int w, int h, COLOR32 Color = CLR_TEXT_DEFAULT);
+	UI_GroupBox(PCHAR Caption, float x, float y, float w, float h, COLOR32 Color = CLR_TEXT_DEFAULT);
 	~UI_GroupBox() { }
 
 public:
-	void Draw(int x, int y, bool Visible = true);
+	void Draw(float x, float y, bool Visible = true);
 
 public:
 	CoreString Text;
-	int X, Y, W, H;
+	float X, Y, W, H;
 	COLOR32 TextColor = CLR_TEXT_DEFAULT;
 	COLOR32 LineColor = CLR_LINE_DEFAULT;
 };
