+++ Framework/Direct3D9/Render/GUI/Controls/UI_Slider.h
@@ -9,17 +9,17 @@ Used for slider handling both vertically and horizontally
 class UI_Slider
 {
 public:
-	UI_Slider(int x, int y, int size, double min, double max, LPCoreValue SliderValue, COLOR32 Color = CLR_SLIDER_DEFAULT, bool DrawVertically = false);
+	UI_Slider(float x, float y, float size, double min, double max, LPCoreValue SliderValue, COLOR32 Color = CLR_SLIDER_DEFAULT, bool DrawVertically = false);
 	~UI_Slider() { }
 
 public:
-	void DrawSliderPin(int x, int y, int size);
-	void Draw(int x, int y, bool Visible = true);
+	void DrawSliderPin(float x, float y, float size);
+	void Draw(float x, float y, bool Visible = true);
 
 public:
 	bool Vertical;
-	int X, Y;
-	int Size;
+	float X, Y;
+	float Size;
 	double Min;
 	double Max;
 	LPCoreValue Value;
